+++ packages/components/histogram/src/index.js
@@ -1,3 +1,3 @@
 export {default as HistogramG} from './HistogramG.svelte';
 export {default as HistogramDiv} from './HistogramDiv.svelte';
-export * from './utils';
+export * from './utils.js';