+++ packages/tools/dom/src/index.js
@@ -1,3 +1,0 @@
-export * from './attrs'
-export * from './html'
-export * from './nodes'