+++ packages/tools/geometry/src/index.js
@@ -1,4 +1,0 @@
-export * from './2DNetwork';
-export * from './2DPoint';
-export * from './2DVector';
-export * from './3D';