+++ packages/tools/request/src/json.js
@@ -9,13 +9,9 @@ import {text} from 'd3-fetch';
 
 /**
  * Return a request to a ndjson file as a promise.
- * Depending on `useFetch` uses the Fetch API or the XMLHttpRequest API.
  *
  * @function
  * @arg {string} url - The URL for the request (GET).
- * @arg {boolean} [useFetch=true] -
- *      Use the Fetch API? (default = `true`).
- *      Useful to pass `Modernizr.fetch` to disable the Fetch API if not available.
  * @return {promise}
  *
  * @example