+++ packages/components/histogram/CHANGELOG.md
@@ -1,3 +1,20 @@
+## `@svizzle/histogram` v0.6.0 (next)
+
+- upgrade to ESM:
+	- `package.json`:
+		- fields:
+			- added `"type": "module"`
+			- added `main`
+			- set `engines.node` to `>=17.5.0`
+		- deps:
+			- added: `d3-array`
+			- upgraded:
+				- `d3-scale`, `eslint`, `eslint-plugin-svelte3`
+			- removed `eslint-plugin-import`, `esm`
+	- imports:
+		- using file extensions
+		- use the `node:` protocol
+
 ## `@svizzle/histogram` v0.5.0
 
 - removed build scripts