+++ packages/components/legend/CHANGELOG.md
@@ -1,3 +1,16 @@
+## `@svizzle/legend` v0.4.0 (next)
+
+- upgrade to ESM:
+	- `package.json`:
+		- fields:
+			- added `"type": "module"`
+			- added `main`
+			- set `engines.node` to `>=17.5.0`
+		- deps:
+			- upgraded:
+				- `eslint`, `eslint-plugin-svelte3`, `yootils`
+			- removed `eslint-plugin-import`
+
 ## `@svizzle/legend` v0.3.0
 
 - removed build scripts